qmail-inject: fix header parse regression
When attempting to catch incorrectly formatted headers like "To: foo: bar, baz"
to avoid errors when injecting bounces the host expansion of headers like
"To: root" was also broken.

Instead of completely ignoring these lines when the recipient list was already
given on the command line using -a option try to process these lines as before,
but allow the parsing of these lines to silently fail. If a broken line is
encountered this now simply does not reformat them to insert the hostname.

Fixes: 56e7c4a74087d47889355c2c55352cc4d0482020